% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{classy3-class}
\alias{classy3-class}
\alias{new_classy3}
\alias{classy_get3}
\alias{classy_count3}
\alias{classy_sum3}
\alias{classy_count_if_less3}
\alias{classy_get_address3}
\title{The \code{classy3} C++ class (uses smart pointers)}
\usage{
new_classy3(x)

classy_get3(p)

classy_count3(p)

classy_sum3(p)

classy_count_if_less3(p, y)

classy_get_address3(p)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{A pointer to a classy3 object.}

\item{y}{Numeric scalar.}
}
\value{
The \code{new_classy3} function creates an object of C++ class \code{classy3}. This is
an object of class \linkS4class{externalptr} in the R side.

\code{classy_sum3} and \code{classy_count3} return the sum and the count of elements
in the \code{x} vector passed to \code{new_classy}.

\code{classy_count_if3} returns the count of elements in \code{x} if these are less
than \code{y}.

\code{classy_get3} returns the original vector \code{x} passed to \code{new_classy}.
}
\description{
The \code{classy3} C++ class (uses smart pointers)
}
\examples{

# Creating a new object
set.seed(1)
obj <- new_classy3(runif(10))

# Getting the sum
classy_sum3(obj)

# Counting
classy_count3(obj)

# Conditional count
classy_count_if_less3(obj, 0.5)

# The object
classy_get3(obj)

# Removing
rm(obj)

# If we call the garbage collector, you can see the destructor been called
gc()
}
