% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_classy}
\alias{new_classy}
\alias{classy}
\alias{classy-class}
\alias{classy_get}
\alias{classy_count}
\alias{classy_sum}
\alias{classy_count_if_less}
\title{The \code{classy} C++ class}
\usage{
new_classy(x_)

classy_get(p_)

classy_count(p_)

classy_sum(p_)

classy_count_if_less(p_, y)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The \code{new_classy} function creates an object of C++ class \code{classy}. This is
an object of class \linkS4class{externalptr} in the R side.

\code{classy_sum} and \code{classy_count} return the sum and the count of elements
in the \code{x} vector passed to \code{new_classy}.

\code{classy_count_if} returns the count of elements in \code{x} if these are less
than \code{y}.

\code{classy_get} returns the original vector \code{x} passed to \code{new_classy}.
}
\description{
The \code{classy} C++ class
}
\examples{

# Creating a new object
set.seed(1)
obj <- new_classy(runif(10))

# Getting the sum
classy_sum(obj)

# Counting
classy_count(obj)

# Conditional count
classy_count_if_less(obj, 0.5)

# The object
classy_get(obj)

# Removing
rm(obj)

# If we call the garbage collector, you can see the destructor been called
gc()
}
