% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_classy1}
\alias{new_classy1}
\alias{classy1}
\alias{classy1-class}
\title{The \code{classy1} C++ class}
\usage{
new_classy1(x_)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The \code{new_classy1} function creates an object of C++ class \code{classy}. This is
an object of class \linkS4class{externalptr} in the R side.

\code{classy_sum1} and \code{classy_count} return the sum and the count of elements
in the \code{x} vector passed to \code{new_classy}.

\code{classy_count_if1} returns the count of elements in \code{x} if these are less
than \code{y}.

\code{classy_get1} returns the original vector \code{x} passed to \code{new_classy}.
}
\description{
The \code{classy1} C++ class
}
\examples{

# Creating a new object
set.seed(1)
obj <- new_classy1(runif(10))

# Getting the sum
classy_sum1(obj)

# Counting
classy_count1(obj)

# Conditional count
classy_count_if_less1(obj, 0.5)

# The object
classy_get1(obj)

# Removing
rm(obj)

# If we call the garbage collector, you can see the destructor been called
gc()
}
