% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{publication}
\alias{publication}
\title{publication}
\usage{
publication(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
returns the product of x and y.
}
\examples{
publication(2,3)
}
